+++ index.html
@@ -7,7 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/main.js" defer></script>
 </head>
 
 <body>
@@ -227,6 +226,6 @@
       </div>
     </section>
   </template>
-  <script src="js/function.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
