+++ js/function.js
@@ -11,10 +11,6 @@ const isPalindromeOne = (string) => string.toLowerCase() === string.toLowerCase(
 
 isPalindromeOne('топот');
 
-// test
-// console.log(isPalindromeOne('топот'));
-// console.log(isPalindromeOne('ДовОд'));
-
 
 // Task_2_2 - Проверка на палиндром строки с пробелами
 
@@ -26,9 +22,6 @@ const isPalindromeTwo = (string) => {
 };
 
 isPalindromeTwo('Лёша на полке клопа нашёл ');
-
-// test
-// console.log(isPalindromeTwo ('Лёша на полке клопа нашёл '));
 
 
 // Task_3_1 - Дополнительное (извлечение цифр из строки)
@@ -50,13 +43,6 @@ const getNumberOne = (string) => {
 };
 
 getNumberOne('1 кефир, 0.5 батона');
-
-// test
-// console.log(getNumberOne('2023 год'));
-// console.log(getNumberOne('ECMAScript 2022'));
-// console.log(getNumberOne('1 кефир, 0.5 батона'));
-// console.log(getNumberOne('агент 007'));
-// console.log(getNumberOne('а я томат'));
 
 
 // Task_3_2 - Дополнительное (извлечение цифр из строки), если пришло число
@@ -84,14 +70,6 @@ const getNumberTwo = (string) => {
 
 getNumberTwo('-1.5');
 
-// test
-// console.log(getNumberTwo('а я томат'));
-// console.log(getNumberTwo('2023 год'));
-// console.log(getNumberTwo(2023));
-// console.log(getNumberTwo(-1.5));
-// console.log(getNumberTwo(0.5));
-
-
 // ВАРИАНТ С РЕГУЛЯРКАМИ
 
 const getNumberThree = (string) => {
@@ -100,7 +78,3 @@ const getNumberThree = (string) => {
 };
 
 getNumberThree('-1.5');
-
-// console.log(getNumberTwo('а я томат'));
-// console.log(getNumberTwo('2023 год'));
-// console.log(getNumberTwo(2023));
